move format conversion functions to new package
