+++ pkg/conversions/conv_test.go
@@ -7,7 +7,7 @@
 // cases as published by the Free Software Foundation.
 //
 
-package utils
+package conversions
 
 import (
 	"testing"
